refactor!: remove ui module
